Add PID filtering

+++ include/ts.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <map>
+#include <stdint.h>
 #include <string>
 
 /**
@@ -9,7 +10,7 @@ namespace ts {
     /**
      * Map standard table names to PID value
      */
-    enum PID : unsigned int {
+    enum PID : uint16_t {
         PAT  = 0x0000,              // Program Association Table
         CAT  = 0x0001,              // Conditional Access Table
         TSDT = 0x0002,              // Transport Stream Description Table
@@ -33,7 +34,7 @@ namespace ts {
     /**
      * Map standard PID values to abbreviated table name
      */
-    static std::map<unsigned int,const char*> PIDMap = {
+    static std::map<uint16_t,const char*> PIDMap = {
         {PID::PAT,  "PAT"},         // Program Association Table
         {PID::CAT,  "CAT"},         // Conditional Access Table
         {PID::TSDT, "TSDT"},        // Transport Stream Description Table
@@ -60,7 +61,7 @@ namespace ts {
         bool            tei;        // Transport Error Indicator Flag
         bool            pusi;       // Payload Unit Start Indicator Flag
         bool            pri;        // Transport Priority Flag
-        unsigned int    pid;        // Packet Identifier
+        uint16_t        pid;        // Packet Identifier
         unsigned int    tsc;        // Transport Scrambling Control
         unsigned int    afc;        // Adaptation Field Control
         unsigned int    cont;       // Continuity Counter
